+++ examples/data-sources/aiven_kafka_user/data-source.tf
@@ -6,6 +6,7 @@ data "aiven_kafka_user" "example" {
   /* COMPUTED FIELDS
   access_cert              = "foo"
   access_key               = "foo"
+  mysql_grants             = ["SELECT", "DELETE"]
   password                 = "password123"
   password_encryption_type = "md5"
   type                     = "foo"