+++ examples/resources/aiven_pg_user/resource.tf
@@ -6,6 +6,7 @@ resource "aiven_pg_user" "example" {
   // OPTIONAL FIELDS
   password_wo          = "password123"
   password_wo_version  = 1
+  mysql_grants         = ["SELECT", "DELETE"] // Force new
   pg_allow_replication = true
 
   /* COMPUTED FIELDS